+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Expose add accounts to a group Admin API [#384](https://github.com/rokwire/core-building-block/issues/384)
+- Expose System APIs for auth types [#362](https://github.com/rokwire/core-building-block/issues/362)
 - Expose grant roles to account Admin API [#383](https://github.com/rokwire/core-building-block/issues/383)
 - Expose grant permissions to account Admin API [#382](https://github.com/rokwire/core-building-block/issues/382)
 - Expose Admin API which gives an application account devices [#359](https://github.com/rokwire/core-building-block/issues/359)
@@ -18,7 +19,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Expose Admin API with the currently logged in accounts [#355](https://github.com/rokwire/core-building-block/issues/355)
 - Add app config endpoints [#261](https://github.com/rokwire/core-building-block/issues/261)
 
+### Security
+- Fix security vulnerability for roles and groups admin APIs [#414](https://github.com/rokwire/core-building-block/issues/414)
+
 ### Fixed
+- Fix issues with account linking [#408](https://github.com/rokwire/core-building-block/issues/408)
 - Fix creating application group admin API [#397](https://github.com/rokwire/core-building-block/issues/397)
 - Fix create role and create group Admin APIs [#386](https://github.com/rokwire/core-building-block/issues/386)
 - Fix broken app config API and CHANGELOG.md [#401](https://github.com/rokwire/core-building-block/issues/401)